BUG: Matrix coordinates not updated after pad.
In GLDZM, image and mask are padded for distance map, however, this requires also an update of the matrixCoordinates to apply the correct mask. This was not provided.

Additionally, by using binImage to make the distancemap discreet, the minimum distance is set to 1 instead of 0. This is accounted for in the C matrix, but not yet in the python matrix (i.e. the index in the matrix should be equal to minDistance - 1, to make it a 0 based index).