ENH: Store abstract functions of filters in dictionary
At initialization, PyRadiomics iterates over the `imageoperations` module to find all possible filters.
As opposed to storing just the names in a list and using `getattr` to find the function upon executing the extraction, store the possible filters in
a dictionary with the filter name as key, and the abstract function as a value.
Then, upon executing the extraction, `getattr` is not needed and the function corresponding to the filter can be called directly from the input image types dictionary.